Fix messed up game state and game files
