+++ src/game.cpp
@@ -1,4 +1,5 @@
 #include "game.h"
+#include "state_identifiers.h"
 #include "utility.h"
 
 #include <iostream>
@@ -6,16 +7,22 @@
 Game::Game()
     : window(sf::VideoMode(1000, 850), "Runaway Monkey",
              sf::Style::Titlebar | sf::Style::Close)
-    , world(window)
+    , textures()
+    , fonts()
     , player()
+    , stateManager(State::Context(window, textures, fonts, player))
     , paused(false)
 {
     //Implicitly use sf::Time::sleep to maintain 60 FPS
     //TODO: Use more reliable method
     window.setFramerateLimit(60);
     window.setKeyRepeatEnabled(false);
-    //Get player actor reference on the player controller object
-    player.setActor(world.getPlayer());
+
+    //TODO: Load titlescreen textures and fonts
+
+    //Register all states and add the titlescreen to the stack
+    registerStates();
+    stateManager.pushState(States::Title);
 }
 
 void Game::run()
@@ -26,9 +33,10 @@ void Game::run()
         //dt = time since the last tick
         sf::Time dt = clock.restart();
         processInput();
-        //Update the game if it's not on pause
-        if (!paused)
-            update(dt);
+        update(dt);
+	//Check if there are no more states left
+	if (stateManager.isEmpty())
+	    window.close();
         render();
     }
 }
@@ -36,13 +44,11 @@ void Game::run()
 //Processes events and real-time input
 void Game::processInput()
 {
-    CommandQueue& commands = world.getCommandQueue();
-
     sf::Event event;
     while (window.pollEvent(event))
     {
         //Pass the event to player
-        player.handleEvent(event, commands);
+        stateManager.handleEvent(event);
 
         //Other actions for events not handled by the player
         if (event.type == sf::Event::Closed)
@@ -52,15 +58,13 @@ void Game::processInput()
         else if (event.type == sf::Event::GainedFocus)
             paused = false;
     }
-
-    //Lets player handle real-time input
-    player.handleRealTimeInput(commands);
 }
 
 //Update the game logic, using dt to make it FPS independent
 void Game::update(sf::Time dt)
 {
-    world.update(dt);
+    if (!paused)
+        stateManager.update(dt);
 }
 
 //Draw everything to buffer and display it on the window
@@ -68,10 +72,17 @@ void Game::render()
 {
     //First clear the window
     window.clear();
-    //Draw everything in the world to buffer
-    world.draw();
+    //Draw all the states in the stack to buffer
+    stateManager.draw();
     //Display buffer
     window.display();
+}
+
+void Game::registerStates()
+{
+    stateManager.registerState<TitleState>(States::Title);
+    stateManager.registerState<MenuState>(States::Menu);
+    stateManager.registerState<GameState>(States::Game);
 }
 
 int main()
